+++ examples/functions-google-gmail/gatsby-config.js
@@ -1,0 +1,8 @@
+require("dotenv").config()
+
+module.exports = {
+  siteMetadata: {
+    title: "Google Auth",
+  },
+  plugins: ["gatsby-plugin-gatsby-cloud"],
+}