+++ packages/gatsby/src/redux/index.ts
@@ -2,6 +2,7 @@ import {
   applyMiddleware,
   combineReducers,
   createStore,
+  DeepPartial,
   Middleware,
   ReducersMapObject,
 } from "redux"
@@ -9,10 +10,10 @@ import _ from "lodash"
 import telemetry from "gatsby-telemetry"
 
 import { mett } from "../utils/mett"
-import thunk, { ThunkMiddleware } from "redux-thunk"
+import thunk, { ThunkMiddleware, ThunkAction } from "redux-thunk"
 import * as reducers from "./reducers"
 import { writeToCache, readFromCache } from "./persist"
-import { IGatsbyState, ActionsUnion } from "./types"
+import { IGatsbyState, ActionsUnion, GatsbyStateKeys } from "./types"
 
 // Create event emitter for actions
 export const emitter = mett()
@@ -61,7 +62,9 @@ export const readState = (): IGatsbyState => {
 }
 
 export interface IMultiDispatch {
-  <T extends ActionsUnion>(action: Array<T>): Array<T>
+  <T extends ActionsUnion | ThunkAction<any, IGatsbyState, any, ActionsUnion>>(
+    action: Array<T>
+  ): Array<T>
 }
 
 /**
@@ -72,7 +75,7 @@ const multi: Middleware<IMultiDispatch> = ({ dispatch }) => next => (
 ): ActionsUnion | Array<ActionsUnion> =>
   Array.isArray(action) ? action.filter(Boolean).map(dispatch) : next(action)
 
-// We're using the inferred type here becauise manually typing it would be very complicated
+// We're using the inferred type here because manually typing it would be very complicated
 // and error-prone. Instead we'll make use of the createStore return value, and export that type.
 // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
 export const configureStore = (initialState: IGatsbyState) =>
@@ -83,7 +86,9 @@ export const configureStore = (initialState: IGatsbyState) =>
   )
 
 export type GatsbyReduxStore = ReturnType<typeof configureStore>
-export const store: GatsbyReduxStore = configureStore(readState())
+export const store: GatsbyReduxStore = configureStore(
+  process.env.GATSBY_WORKER_POOL_WORKER ? ({} as IGatsbyState) : readState()
+)
 
 /**
  * Allows overloading some reducers (e.g. when setting a custom datastore)
@@ -91,7 +96,9 @@ export const store: GatsbyReduxStore = configureStore(readState())
 export function replaceReducer(
   customReducers: Partial<ReducersMapObject<IGatsbyState>>
 ): void {
-  store.replaceReducer(combineReducers({ ...reducers, ...customReducers }))
+  store.replaceReducer(
+    combineReducers<IGatsbyState>({ ...reducers, ...customReducers })
+  )
 }
 
 // Persist state.
@@ -119,6 +126,24 @@ export const saveState = (): void => {
     queries: state.queries,
     html: state.html,
   })
+}
+
+export const saveStateForWorkers = (slices: Array<GatsbyStateKeys>): void => {
+  const state = store.getState()
+  const contents = _.pick(state, slices)
+
+  return writeToCache(contents, slices)
+}
+
+export const loadStateInWorker = (
+  slices: Array<GatsbyStateKeys>
+): DeepPartial<IGatsbyState> => {
+  try {
+    return readFromCache(slices) as DeepPartial<IGatsbyState>
+  } catch (e) {
+    // ignore errors.
+  }
+  return {} as IGatsbyState
 }
 
 store.subscribe(() => {